Mysql/requirement.txt
pymysql 
fastmcp
